Changed from try catch to if

+++ training/checkpointsaver.py
@@ -23,10 +23,8 @@ class SaveBestNCheckpoints(tf.keras.callbacks.ModelCheckpoint):
                 # If we already have N checkpoints, we delete the worst one to make room for the new one
                 if len(self._checkpoints) >= self.n:
                     removed_checkpoint = self._checkpoints.pop(-1)
-                    try:
+                    if os.path.exists(removed_checkpoint["path"]):
                         os.remove(removed_checkpoint["path"])
-                    except FileNotFoundError:
-                        pass
 
 
                 # We add the new checkpoint to the list
